Fixed inconsistency in interest calculation between planned payment and payment by accruing interest on the real amount rather than compounding the calculated daily interest, then applying it for the planned payment.
This resulted in some misbehavior in planned payments around the interest
costcomponent since both accrual and incurral were being added to that
costcomponent.  To fix this I excluded accrued costcomponents from the
payment and changed the basis of journal entry creation from the
costcomponents to the balanceAdjustments in the payment builder.

This approach gives a more accurate result anyways because sometimes
when two identical charges were debitting the same account, one of them
was getting lost in the set.  It also makes it easier to filter out
entries which cancel each other producing a zero debit or credit.